+++ clang-tools-extra/clang-tidy/readability/UniqueptrDeleteReleaseCheck.cpp
@@ -9,6 +9,8 @@
 #include "UniqueptrDeleteReleaseCheck.h"
 #include "clang/AST/ASTContext.h"
 #include "clang/ASTMatchers/ASTMatchFinder.h"
+#include "clang/Basic/Diagnostic.h"
+#include "clang/Basic/SourceLocation.h"
 #include "clang/Lex/Lexer.h"
 
 using namespace clang::ast_matchers;
@@ -17,50 +19,69 @@ namespace clang {
 namespace tidy {
 namespace readability {
 
+void UniqueptrDeleteReleaseCheck::storeOptions(
+    ClangTidyOptions::OptionMap &Opts) {
+  Options.store(Opts, "PreferResetCall", PreferResetCall);
+}
+
+UniqueptrDeleteReleaseCheck::UniqueptrDeleteReleaseCheck(
+    StringRef Name, ClangTidyContext *Context)
+    : ClangTidyCheck(Name, Context),
+      PreferResetCall(Options.get("PreferResetCall", false)) {}
+
 void UniqueptrDeleteReleaseCheck::registerMatchers(MatchFinder *Finder) {
-  auto IsSusbstituted = qualType(anyOf(
-      substTemplateTypeParmType(), hasDescendant(substTemplateTypeParmType())));
 
   auto UniquePtrWithDefaultDelete = classTemplateSpecializationDecl(
-      hasName("std::unique_ptr"),
-      hasTemplateArgument(1, refersToType(qualType(hasDeclaration(cxxRecordDecl(
-                                 hasName("std::default_delete")))))));
+      hasName("::std::unique_ptr"),
+      hasTemplateArgument(1, refersToType(hasDeclaration(cxxRecordDecl(
+                                 hasName("::std::default_delete"))))));
 
   Finder->addMatcher(
-      cxxDeleteExpr(has(ignoringParenImpCasts(cxxMemberCallExpr(
-                        on(expr(hasType(UniquePtrWithDefaultDelete),
-                                unless(hasType(IsSusbstituted)))
-                               .bind("uptr")),
-                        callee(cxxMethodDecl(hasName("release")))))))
+      cxxDeleteExpr(
+          unless(isInTemplateInstantiation()),
+          has(expr(ignoringParenImpCasts(
+              cxxMemberCallExpr(
+                  callee(
+                      memberExpr(hasObjectExpression(allOf(
+                                     unless(isTypeDependent()),
+                                     anyOf(hasType(UniquePtrWithDefaultDelete),
+                                           hasType(pointsTo(
+                                               UniquePtrWithDefaultDelete))))),
+                                 member(cxxMethodDecl(hasName("release"))))
+                          .bind("release_expr")))
+                  .bind("release_call")))))
           .bind("delete"),
       this);
 }
 
 void UniqueptrDeleteReleaseCheck::check(
     const MatchFinder::MatchResult &Result) {
-  const auto *PtrExpr = Result.Nodes.getNodeAs<Expr>("uptr");
-  const auto *DeleteExpr = Result.Nodes.getNodeAs<Expr>("delete");
+  const auto *DeleteExpr = Result.Nodes.getNodeAs<CXXDeleteExpr>("delete");
+  const auto *ReleaseExpr = Result.Nodes.getNodeAs<MemberExpr>("release_expr");
+  const auto *ReleaseCallExpr =
+      Result.Nodes.getNodeAs<CXXMemberCallExpr>("release_call");
 
-  if (PtrExpr->getBeginLoc().isMacroID())
+  if (ReleaseExpr->getBeginLoc().isMacroID())
     return;
 
-  // Ignore dependent types.
-  // It can give us false positives, so we go with false negatives instead to
-  // be safe.
-  if (PtrExpr->getType()->isDependentType())
-    return;
-
-  SourceLocation AfterPtr = Lexer::getLocForEndOfToken(
-      PtrExpr->getEndLoc(), 0, *Result.SourceManager, getLangOpts());
-
-  diag(DeleteExpr->getBeginLoc(),
-       "prefer '= nullptr' to 'delete x.release()' to reset unique_ptr<> "
-       "objects")
-      << FixItHint::CreateRemoval(CharSourceRange::getCharRange(
-             DeleteExpr->getBeginLoc(), PtrExpr->getBeginLoc()))
-      << FixItHint::CreateReplacement(
-             CharSourceRange::getTokenRange(AfterPtr, DeleteExpr->getEndLoc()),
-             " = nullptr");
+  auto D =
+      diag(DeleteExpr->getBeginLoc(), "prefer '%select{= nullptr|reset()}0' "
+                                      "to reset 'unique_ptr<>' objects");
+  D << PreferResetCall << DeleteExpr->getSourceRange()
+    << FixItHint::CreateRemoval(CharSourceRange::getCharRange(
+           DeleteExpr->getBeginLoc(),
+           DeleteExpr->getArgument()->getBeginLoc()));
+  if (PreferResetCall) {
+    D << FixItHint::CreateReplacement(ReleaseExpr->getMemberLoc(), "reset");
+  } else {
+    if (ReleaseExpr->isArrow())
+      D << FixItHint::CreateInsertion(ReleaseExpr->getBase()->getBeginLoc(),
+                                      "*");
+    D << FixItHint::CreateReplacement(
+        CharSourceRange::getTokenRange(ReleaseExpr->getOperatorLoc(),
+                                       ReleaseCallExpr->getEndLoc()),
+        " = nullptr");
+  }
 }
 
 } // namespace readability